+++ src/RcsbFvSequence/SequenceViews/AssemblyView/AssemblyView.tsx
@@ -2,10 +2,7 @@ import {asyncScheduler} from "rxjs";
 import * as React from "react";
 
 import {RcsbFvDOMConstants} from "../../../RcsbFvConstants/RcsbFvConstants";
-import {
-    buildInstanceSequenceFv,
-    unmount
-} from "@rcsb/rcsb-saguaro-app";
+import {buildInstanceSequenceFv, FeatureType, RcsbFvContextManager, RcsbFvUI, unmount} from "@rcsb/rcsb-saguaro-app";
 import {AbstractView, AbstractViewInterface} from "../AbstractView";
 import {
     InstanceSequenceConfig,
@@ -14,8 +11,11 @@ import {
 import {RcsbFvBoardConfigInterface, RcsbFvTrackDataElementInterface} from "@rcsb/rcsb-saguaro";
 import {
     ChainInfo,
+    OperatorInfo,
     SaguaroPluginInterface,
-    SaguaroPluginModelMapType, SaguaroRange, SaguaroRegionList
+    SaguaroPluginModelMapType,
+    SaguaroRange,
+    SaguaroRegionList
 } from "../../../RcsbFvStructure/SaguaroPluginInterface";
 import {OptionPropsInterface} from "@rcsb/rcsb-saguaro-app/build/dist/RcsbFvWeb/WebTools/SelectButton";
 
@@ -27,28 +27,27 @@ import {
     RcsbFvAdditionalConfig,
     RcsbFvModulePublicInterface
 } from "@rcsb/rcsb-saguaro-app/build/dist/RcsbFvWeb/RcsbFvModule/RcsbFvModuleInterface";
-import {RcsbFvUI} from "@rcsb/rcsb-saguaro-app";
+import {AnnotationFeatures, Source, Type} from "@rcsb/rcsb-api-tools/build/RcsbGraphQL/Types/Borrego/GqlTypes";
+import {PolymerEntityInstanceInterface} from "@rcsb/rcsb-saguaro-app/build/dist/RcsbCollectTools/Translators/PolymerEntityInstancesCollector";
+import {InterfaceInstanceTranslate} from "@rcsb/rcsb-saguaro-app/build/dist/RcsbUtils/Translators/InterfaceInstanceTranslate";
+import {AssemblyModelSate} from "./AssemblyModelSate";
 
 export interface AssemblyViewInterface {
     entryId: string;
-    additionalConfig?: RcsbFvAdditionalConfig;
+    additionalConfig?: RcsbFvAdditionalConfig & {operatorChangeCallback?:(operatorInfo: OperatorInfo)=>void};
     instanceSequenceConfig?: InstanceSequenceConfig;
     useOperatorsFlag?:boolean;
 }
 
 export class AssemblyView extends AbstractView<AssemblyViewInterface & AbstractViewInterface, {}>{
 
-    private currentLabelAsymId: string;
-    private currentEntryId: string;
-    private currentModelId: string;
-    private currentOpName: string | undefined = undefined;
-    private currentModelNumber: string;
+    private readonly assemblyModelSate: AssemblyModelSate = new AssemblyModelSate();
     private createComponentThreshold: number = 3;
     private innerSelectionFlag: boolean = false;
     private currentSelectedComponentId: string;
-    private currentModelMap:SaguaroPluginModelMapType;
     private boardConfig: Partial<RcsbFvBoardConfigInterface>;
     private rcsbFvModule: RcsbFvModulePublicInterface | null;
+    private OPERATOR_DROPDOWN_TITLE: string = "Symmetry Partner";
     //private readonly componentSet = new Map<string, {current: Set<string>, previous: Set<string>}>();
 
     additionalContent(): JSX.Element {
@@ -144,15 +143,30 @@ export class AssemblyView extends AbstractView<AssemblyViewInterface & AbstractV
             this.rcsbFvModule?.getFv().clearSelection(mode);
             if(mode === 'select')
                 this.resetPluginView();
-        }else if(mode === 'select' && this.props.selectorManager.getLastSelection('select')?.labelAsymId != null && this.props.selectorManager.getLastSelection('select')?.labelAsymId != this.currentLabelAsymId){
-            const authId: string | undefined = this.currentModelMap
-                .get(this.currentModelId)?.chains
-                .filter(ch=>(ch.label===this.props.selectorManager.getLastSelection('select')?.labelAsymId))[0]?.auth;
-            await this.modelChangeCallback(this.currentModelMap, authId, this.props.selectorManager.getLastSelection('select')?.operatorName);
+        }else if(
+            mode === 'select' &&
+            this.props.selectorManager.getLastSelection('select')?.labelAsymId != null &&
+            this.props.selectorManager.getLastSelection('select')?.labelAsymId != this.assemblyModelSate.getString("labelAsymId")
+        ){
+            const authId: string | undefined = this.assemblyModelSate.getChainInfo(this.props.selectorManager.getLastSelection('select')?.labelAsymId!)?.auth;
+            await this.modelChangeCallback(this.assemblyModelSate.getMap(), authId, this.props.selectorManager.getLastSelection('select')?.operatorName);
+        }else if(
+            mode === 'select' &&
+            this.props.selectorManager.getLastSelection('select')?.labelAsymId != null &&
+            this.props.selectorManager.getLastSelection('select')?.operatorName != null &&
+            this.props.selectorManager.getLastSelection('select')?.operatorName != this.assemblyModelSate.getOperator()?.name
+        ){
+            const authId: string | undefined = this.assemblyModelSate.getChainInfo(this.props.selectorManager.getLastSelection('select')?.labelAsymId!)?.auth;
+            await this.modelChangeCallback(this.assemblyModelSate.getMap(), authId, this.props.selectorManager.getLastSelection('select')?.operatorName);
         }else{
-            if(mode === 'select' && this.props.selectorManager.getLastSelection('select')?.operatorName && this.props.selectorManager.getLastSelection('select')?.operatorName != this.currentOpName)
+            if(mode === 'select' && this.props.selectorManager.getLastSelection('select')?.operatorName && this.props.selectorManager.getLastSelection('select')?.operatorName != this.assemblyModelSate.getOperator()?.name)
                 this.addOperatorButton(this.props.selectorManager.getLastSelection('select')?.operatorName);
-            const sel: SaguaroRegionList | undefined = this.props.selectorManager.getSelectionWithCondition(this.currentModelId, this.currentLabelAsymId, mode, this.currentOpName);
+            const sel: SaguaroRegionList | undefined = this.props.selectorManager.getSelectionWithCondition(
+                this.assemblyModelSate.getString("modelId"),
+                this.assemblyModelSate.getString("labelAsymId"),
+                mode,
+                this.assemblyModelSate.getOperator()?.name
+            );
             if (sel == null) {
                 this.rcsbFvModule?.getFv().clearSelection(mode);
                 if(mode === 'select')
@@ -165,36 +179,47 @@ export class AssemblyView extends AbstractView<AssemblyViewInterface & AbstractV
     }
 
     async modelChangeCallback(modelMap:SaguaroPluginModelMapType, defaultAuthId?: string, defaultOperatorName?:string): Promise<void> {
-        this.currentModelMap = modelMap;
-        this.currentOpName = undefined;
+        this.assemblyModelSate.setMap(modelMap);
         this.props.plugin.clearFocus();
         const onChangeCallback: Map<string, (x: InstanceSequenceOnchangeInterface)=>void> = new Map<string, (x: InstanceSequenceOnchangeInterface) => {}>();
-        const filterInstances: Map<string, Set<string>> = new Map<string, Set<string>>();
-        modelMap.forEach((v,k)=>{
+        const assemblyInstances: Map<string, Set<string>> = new Map<string, Set<string>>();
+        this.assemblyModelSate.forEach((v,k)=>{
+            assemblyInstances.set(v.entryId,new Set<string>(v.chains.map(d=>d.label)));
             onChangeCallback.set(v.entryId,(x)=>{
-                this.currentEntryId = v.entryId;
-                this.currentLabelAsymId = x.asymId;
-                this.currentModelId = k;
+                this.assemblyModelSate.set({entryId: v.entryId, labelAsymId: x.asymId, modelId: k});
                 asyncScheduler.schedule(()=>{
                     this.props.selectorManager.setLastSelection('select', null);
                     this.instanceChangeCallback();
                 },1000);
-                this.addOperatorButton(defaultOperatorName);
             });
-            filterInstances.set(v.entryId,new Set<string>(v.chains.map(d=>d.label)));
         });
         this.unmountRcsbFv();
-        const entryId: string = Array.from(modelMap.values()).map(d=>d.entryId)[0];
-        if(entryId != null) {
+        const operatorNameContainer: {operatorName?:string} = {operatorName: defaultOperatorName};
+        if(this.assemblyModelSate.get("entryId") != null) {
             this.rcsbFvModule = await buildInstanceSequenceFv(
                 this.rcsbFvDivId,
                 RcsbFvDOMConstants.SELECT_INSTANCE_PFV_ID,
-                entryId,
+                this.assemblyModelSate.getString("entryId"),
                 {
                     ...this.props.instanceSequenceConfig,
                     defaultValue: defaultAuthId,
-                    onChangeCallback: onChangeCallback.get(entryId),
-                    filterInstances: filterInstances.get(entryId),
+                    onChangeCallback: onChangeCallback.get(this.assemblyModelSate.getString("entryId")),
+                    beforeChangeCallback: (x: InstanceSequenceOnchangeInterface)=>{
+                        this.assemblyModelSate.set({entryId:x.pdbId, labelAsymId: x.asymId});
+                        //TODO this will only work when modelId is equal to pdbId
+                        const operator: OperatorInfo|undefined = getOperator(this.assemblyModelSate.getMap().get(x.pdbId)!, defaultAuthId, operatorNameContainer.operatorName);
+                        this.addOperatorButton(operator?.name);
+                        this.assemblyModelSate.setOperator(x.asymId,operator?.name);
+                        operatorNameContainer.operatorName = undefined;
+                        if(typeof this.props.additionalConfig?.operatorChangeCallback === "function" && this.assemblyModelSate.getOperator()){
+                                this.props.additionalConfig.operatorChangeCallback(this.assemblyModelSate.getOperator()!);
+                        }
+                        if((this.assemblyModelSate.getChainInfo()?.operators?.length ?? 0) > 1)
+                            return {
+                                operatorIds: operator?.ids
+                            }
+                    },
+                    filterInstances: assemblyInstances.get(this.assemblyModelSate.getString("entryId")),
                     selectButtonOptionProps: (props: OptionProps<OptionPropsInterface>) => (components.Option &&
                         <div style={{display: 'flex'}}>
                             <ChainDisplay plugin={this.props.plugin} label={props.data.label}/>
@@ -203,12 +228,15 @@ export class AssemblyView extends AbstractView<AssemblyViewInterface & AbstractV
                 },
                 {
                     ...this.props.additionalConfig,
-                    boardConfig: this.boardConfig
+                    boardConfig: this.boardConfig,
+                    externalTrackBuilder:{
+                        filterFeatures: this.filterFeatures.bind(this)
+                    }
                 }
             );
         }
         if(!defaultAuthId)
-            await createComponents(this.props.plugin, modelMap);
+            await createComponents(this.props.plugin, this.assemblyModelSate.getMap());
     }
 
     private async instanceChangeCallback(): Promise<void>{
@@ -216,26 +244,28 @@ export class AssemblyView extends AbstractView<AssemblyViewInterface & AbstractV
         await this.pluginSelectCallback('select');
     }
 
-    private addOperatorButton(operatorName?:string): void{
-        const currentChainInfo: ChainInfo|undefined = this.currentModelMap.get(this.currentModelId)?.chains.find(ch=>ch.label===this.currentLabelAsymId);
+    private addOperatorButton(operatorName?: string): void{
+        const currentChainInfo: ChainInfo|undefined = this.assemblyModelSate.getChainInfo();
         if(this.props.useOperatorsFlag && currentChainInfo && currentChainInfo.operators.length >1 ){
-            this.currentOpName = operatorName ?? currentChainInfo.operators[0].name;
+            this.assemblyModelSate.setOperator(undefined,operatorName);
             RcsbFvUI.addSelectButton(
                 this.rcsbFvDivId,
                 RcsbFvDOMConstants.SELECT_INSTANCE_PFV_ID,
                 currentChainInfo.operators.map(op=>({
-                    label:op.name,
+                    label:`${op.ids.join("-")} (${op.name})`,
                     optId:op.name,
-                    onChange:()=>{
-                        this.currentOpName = op.name;
-                        asyncScheduler.schedule(()=>{
-                            this.props.selectorManager.setLastSelection('select', null);
-                            this.structureSelectionCallback();
-                        },300);
+                    onChange: async ()=>{
+                        this.assemblyModelSate.set({operator:op});
+                        await this.modelChangeCallback(
+                            this.assemblyModelSate.getMap(),
+                            this.assemblyModelSate.getChainInfo()?.auth,
+                            op.name
+                        )
                     }
                 })),
                 {
-                    defaultValue: this.currentOpName
+                    defaultValue: this.assemblyModelSate.getOperator()?.name,
+                    dropdownTitle:this.OPERATOR_DROPDOWN_TITLE
                 }
             );
         }
@@ -249,15 +279,26 @@ export class AssemblyView extends AbstractView<AssemblyViewInterface & AbstractV
     private highlightHoverCallback(selection: RcsbFvTrackDataElementInterface[]): void {
         if(selection != null && selection.length > 0) {
             if(selection[0].isEmpty){
-                const selectionList = [{modelId: this.currentModelId, labelAsymId: this.currentLabelAsymId, position: selection[0].begin, operatorName: this.currentOpName}];
-                if(selection[0].end != null) selectionList.push({modelId: this.currentModelId, labelAsymId: this.currentLabelAsymId, position: selection[0].end, operatorName: this.currentOpName})
+                const selectionList = [{
+                    modelId: this.assemblyModelSate.getString("modelId"),
+                    labelAsymId: this.assemblyModelSate.getString("labelAsymId"),
+                    position: selection[0].begin,
+                    operatorName: this.assemblyModelSate.getOperator()?.name
+                }];
+                if(selection[0].end != null)
+                    selectionList.push({
+                        modelId: this.assemblyModelSate.getString("modelId"),
+                        labelAsymId: this.assemblyModelSate.getString("labelAsymId"),
+                        position: selection[0].end,
+                        operatorName: this.assemblyModelSate.getOperator()?.name
+                    })
                 this.props.plugin.select(
                     selectionList,
                     'hover',
                     'set'
                 );
             }else {
-                this.props.plugin.select(processMultipleGaps(this.currentModelId, this.currentLabelAsymId, selection, this.currentOpName), 'hover', 'set');
+                this.props.plugin.select(processMultipleGaps(this.assemblyModelSate.getString("modelId"), this.assemblyModelSate.getString("labelAsymId"), selection, this.assemblyModelSate.getOperator()?.name), 'hover', 'set');
             }
         }else{
             this.props.plugin.clearSelection('hover');
@@ -267,8 +308,8 @@ export class AssemblyView extends AbstractView<AssemblyViewInterface & AbstractV
     private selectionChangeCallback(selection: Array<RcsbFvTrackDataElementInterface>): void {
         if(this.innerSelectionFlag)
             return;
-        this.props.plugin.clearSelection('select', {modelId: this.currentModelId, labelAsymId: this.currentLabelAsymId, operatorName: this.currentOpName});
-        this.props.selectorManager.clearSelection('select', {labelAsymId: this.currentLabelAsymId, operatorName: this.currentOpName});
+        this.props.plugin.clearSelection('select', {modelId: this.assemblyModelSate.getString("modelId"), labelAsymId: this.assemblyModelSate.getString("labelAsymId"), operatorName: this.assemblyModelSate.getOperator()?.name});
+        this.props.selectorManager.clearSelection('select', {labelAsymId: this.assemblyModelSate.getString("labelAsymId"), operatorName: this.assemblyModelSate.getOperator()?.name});
         if(selection == null || selection.length === 0) {
             this.resetPluginView();
         }else{
@@ -281,14 +322,28 @@ export class AssemblyView extends AbstractView<AssemblyViewInterface & AbstractV
             const x = e.begin;
             const y = e.end ?? e.begin;
             if(e.isEmpty){
-                this.props.plugin.select(
-                    [{modelId: this.currentModelId, labelAsymId: this.currentLabelAsymId, position: x, operatorName: this.currentOpName},{modelId: this.currentModelId, labelAsymId: this.currentLabelAsymId, position: y, operatorName: this.currentOpName}], 'select',
-                    'add'
+                this.props.plugin.select([{
+                        modelId: this.assemblyModelSate.getString("modelId"),
+                        labelAsymId: this.assemblyModelSate.getString("labelAsymId"),
+                        position: x,
+                        operatorName: this.assemblyModelSate.getOperator()?.name},
+                    {
+                        modelId: this.assemblyModelSate.getString("modelId"),
+                        labelAsymId: this.assemblyModelSate.getString("labelAsymId"),
+                        position: y,
+                        operatorName: this.assemblyModelSate.getOperator()?.name
+                    }],
+                    'select',
+                 'add'
                 );
-                this.props.selectorManager.addSelectionFromRegion(this.currentModelId, this.currentLabelAsymId, {begin:x, end:y, isEmpty: true, source: 'sequence'}, 'select', this.currentOpName);
+                this.props.selectorManager.addSelectionFromRegion(
+                    this.assemblyModelSate.getString("modelId"),
+                    this.assemblyModelSate.getString("labelAsymId"),
+                    {begin:x, end:y, isEmpty: true, source: 'sequence'},
+                    'select', this.assemblyModelSate.getOperator()?.name);
             }else{
-                this.props.plugin.select(processGaps(this.currentModelId, this.currentLabelAsymId, e, this.currentOpName), 'select', 'add');
-                this.props.selectorManager.addSelectionFromRegion(this.currentModelId, this.currentLabelAsymId, {begin:x, end:y, source: 'sequence'}, 'select', this.currentOpName);
+                this.props.plugin.select(processGaps(this.assemblyModelSate.getString("modelId"), this.assemblyModelSate.getString("labelAsymId"), e, this.assemblyModelSate.getOperator()?.name), 'select', 'add');
+                this.props.selectorManager.addSelectionFromRegion(this.assemblyModelSate.getString("modelId"), this.assemblyModelSate.getString("labelAsymId"), {begin:x, end:y, source: 'sequence'}, 'select', this.assemblyModelSate.getOperator()?.name);
             }
         });
     }
@@ -302,40 +357,68 @@ export class AssemblyView extends AbstractView<AssemblyViewInterface & AbstractV
         const x = e.begin;
         const y = e.end ?? e.begin;
         if(e.isEmpty){
-            this.props.plugin.cameraFocus(this.currentModelId, this.currentLabelAsymId, [x,y], this.currentOpName);
-            this.currentSelectedComponentId = this.currentLabelAsymId +":"+ ((x === y) ? x.toString() : x.toString()+","+y.toString());
+            this.props.plugin.cameraFocus(this.assemblyModelSate.getString("modelId"), this.assemblyModelSate.getString("labelAsymId"), [x,y], this.assemblyModelSate.getOperator()?.name);
+            this.currentSelectedComponentId = this.assemblyModelSate.getString("labelAsymId") +":"+ ((x === y) ? x.toString() : x.toString()+","+y.toString());
             asyncScheduler.schedule(async ()=>{
                 await this.props.plugin.createComponent(
                     this.currentSelectedComponentId,
                     [
-                        {modelId: this.currentModelId, labelAsymId: this.currentLabelAsymId, position: x, operatorName: this.currentOpName},
-                        {modelId: this.currentModelId, labelAsymId: this.currentLabelAsymId, position: y, operatorName: this.currentOpName}
+                        {modelId: this.assemblyModelSate.getString("modelId"), labelAsymId: this.assemblyModelSate.getString("labelAsymId"), position: x, operatorName: this.assemblyModelSate.getOperator()?.name},
+                        {modelId: this.assemblyModelSate.getString("modelId"), labelAsymId: this.assemblyModelSate.getString("labelAsymId"), position: y, operatorName: this.assemblyModelSate.getOperator()?.name}
                         ],
                     'ball-and-stick'
                 )
                 if(x === y)
                     asyncScheduler.schedule(()=>{
-                        this.props.plugin.setFocus(this.currentModelId, this.currentLabelAsymId, x, y, this.currentOpName);
+                        this.props.plugin.setFocus(this.assemblyModelSate.getString("modelId"), this.assemblyModelSate.getString("labelAsymId"), x, y, this.assemblyModelSate.getOperator()?.name);
                     },200);
             },100);
 
         }else{
-            this.props.plugin.cameraFocus(this.currentModelId, this.currentLabelAsymId, x, y, this.currentOpName);
+            this.props.plugin.cameraFocus(this.assemblyModelSate.getString("modelId"), this.assemblyModelSate.getString("labelAsymId"), x, y, this.assemblyModelSate.getOperator()?.name);
             if((y-x)<this.createComponentThreshold){
-                this.currentSelectedComponentId = this.currentLabelAsymId +":"+ (x === y ? x.toString() : x.toString()+"-"+y.toString());
+                this.currentSelectedComponentId = this.assemblyModelSate.getString("labelAsymId") +":"+ (x === y ? x.toString() : x.toString()+"-"+y.toString());
                 asyncScheduler.schedule(async ()=>{
                     await this.props.plugin.createComponent(
                         this.currentSelectedComponentId,
-                        processGaps(this.currentModelId, this.currentLabelAsymId, e, this.currentOpName),
+                        processGaps(this.assemblyModelSate.getString("modelId"), this.assemblyModelSate.getString("labelAsymId"), e, this.assemblyModelSate.getOperator()?.name),
                         'ball-and-stick'
                     )
                     if(x === y)
                         asyncScheduler.schedule(()=>{
-                            this.props.plugin.setFocus(this.currentModelId, this.currentLabelAsymId, x, y, this.currentOpName);
+                            this.props.plugin.setFocus(this.assemblyModelSate.getString("modelId"), this.assemblyModelSate.getString("labelAsymId"), x, y, this.assemblyModelSate.getOperator()?.name);
                         },200);
                 },100);
             }
         }
+    }
+
+    private filterFeatures(data: {annotations: Array<AnnotationFeatures>; rcsbContext:Partial<PolymerEntityInstanceInterface>}): Promise<Array<AnnotationFeatures>> {
+        return new Promise<Array<AnnotationFeatures>>(async resolve => {
+            let annotations: Array<AnnotationFeatures> = [];
+            (await Promise.all(data.annotations.map(async ann=>{
+                if(ann.source == Source.PdbInterface && ann.target_id && data.rcsbContext?.asymId) {
+                    const interfaceToInstance: InterfaceInstanceTranslate = await RcsbFvContextManager.getInterfaceToInstance(ann.target_id);
+                    if(typeof ann.target_identifiers?.interface_partner_index === "number" && ann.target_identifiers.assembly_id === this.assemblyModelSate.getString("assemblyId")) {
+                        const operatorIds:string[][] = interfaceToInstance.getOperatorIds(ann.target_id)[ann.target_identifiers.interface_partner_index];
+                        if(ann.features && this.assemblyModelSate.getOperator() && operatorIds.map(o=>o.join("|")).includes( this.assemblyModelSate.getOperator()!.ids.join("|") )){
+                            ann.features = ann.features.filter(f=>(f && f.type == FeatureType.BurialFraction));
+                            if(ann.features.length > 0)
+                                return ann;
+                        }
+                    }
+                }else if(ann.source == Source.PdbInstance && ann.features){
+                    ann.features = ann.features?.filter(f=>(f?.type!==Type.Asa));
+                    return ann;
+                }else if(ann.source != Source.PdbInterface){
+                    return ann;
+                }
+            }))).forEach((value,index,array)=>{
+                if(value)
+                    annotations = annotations.concat(value);
+            });
+            resolve(annotations);
+        });
     }
 
 }
@@ -390,4 +473,14 @@ async function createComponents(plugin: SaguaroPluginInterface, modelMap:Saguaro
         await plugin.colorComponent(label, 'chain-id');
     }
     await plugin.removeComponent("Polymer");
+}
+
+function getOperator(entryInfo: {entryId: string; assemblyId: string, chains:Array<ChainInfo>;}, defaultAuthId?: string, defaultOperatorName?:string): OperatorInfo | undefined{
+    const chainInfo: ChainInfo | undefined = defaultAuthId ? entryInfo.chains.find(ch=>ch.auth === defaultAuthId) : entryInfo.chains[0];
+    if(chainInfo){
+        const operatorInfo: OperatorInfo | undefined = defaultOperatorName ? chainInfo.operators.find(op=>op.name === defaultOperatorName) : chainInfo.operators[0];
+        if(operatorInfo)
+            return operatorInfo;
+    }
+    return undefined;
 }